lab3/livelab/trigger-vertext-ai/requirements.txt
google-cloud-aiplatform[pipelines]
google-cloud-storage